fix(mobile): stop unrenewed-grace rotation churn and gate cadence gaps
- skip proactive rotation when the resume confirmation reports renewed=false
  (a re-resume provably returns the same unchanged deadline; rotating churned
  one session replacement per clamp floor, ~60/hour, until a fresh credential)
- armCredentialReprobe under a held gate mints the tick's pass token so the
  effective reprobe cadence stays 60s..15min instead of doubling to ~30min
- registerFailure honors scheduleRetry=false in gate branches: no reprobe
  timer is armed while backgrounded/stopped; foreground resume re-arms
- extract RelayRetryDelays and supervisor test fakes into their own modules
  (max-lines)
